requirements.txt
discord
asyncio
pillow
requests
pyfiglet
lyricsgenius
youtube_dl
selenium
googletrans
google-api-python-client